+++ tests/php/Unit/Compiler/Config/PhelConfigTest.php
@@ -11,7 +11,36 @@ use PHPUnit\Framework\TestCase;
 
 final class PhelConfigTest extends TestCase
 {
-    public function test_json_serialize(): void
+    public function test_default_json_serialize(): void
+    {
+        $config = new PhelConfig();
+
+        $expected = [
+            PhelConfig::SRC_DIRS => ['src'],
+            PhelConfig::TEST_DIRS => ['tests'],
+            PhelConfig::VENDOR_DIR => 'vendor',
+            PhelConfig::ERROR_LOG_FILE => 'data/error.log',
+            PhelConfig::BUILD_CONFIG => [
+                PhelBuildConfig::MAIN_PHEL_NAMESPACE => '',
+                PhelBuildConfig::DEST_DIR => 'out',
+                PhelBuildConfig::MAIN_PHP_FILENAME => 'index.php',
+                PhelBuildConfig::MAIN_PHP_PATH => 'out/index.php',
+            ],
+            PhelConfig::EXPORT_CONFIG => [
+                PhelExportConfig::TARGET_DIRECTORY => 'src/PhelGenerated',
+                PhelExportConfig::FROM_DIRECTORIES => ['src'],
+                PhelExportConfig::NAMESPACE_PREFIX => 'PhelGenerated',
+            ],
+            PhelConfig::IGNORE_WHEN_BUILDING => ['ignore-when-building.phel'],
+            PhelConfig::NO_CACHE_WHEN_BUILDING => [],
+            PhelConfig::KEEP_GENERATED_TEMP_FILES => false,
+            PhelConfig::FORMAT_DIRS => ['src', 'tests'],
+        ];
+
+        self::assertSame($expected, $config->jsonSerialize());
+    }
+
+    public function test_custom_json_serialize(): void
     {
         $config = (new PhelConfig())
             ->setSrcDirs(['some/directory'])
@@ -20,7 +49,7 @@ final class PhelConfigTest extends TestCase
             ->setErrorLogFile('error-log.file')
             ->setBuildConfig((new PhelBuildConfig())
                 ->setMainPhpPath('out/custom-index.php')
-                ->setMainPhelNamespace('test-ns/boot'), )
+                ->setMainPhelNamespace('test-ns/boot'))
             ->setExportConfig((new PhelExportConfig())
                 ->setFromDirectories(['some/other/dir'])
                 ->setNamespacePrefix('Generated')